Remove non-compact XML schema, don't validate in Python code
The schema is already checked using the RelaxNG-compact definition,
which is the authoritative version anyway.  Checking in the Python
code again is unnecessary, as the source is already checked by the
make check target.

All in all, this commit reduces the number of (identical!)  checks
from three to one during build time.