Corrected paper title (#629)
* Corrected paper title

Paper title of P10-4004 is given as "GernEdiT: A Graphical Tool for GermaNet Development" in the PDF, but metadata and proceedings TOC give it as "GernEdiT - The GermaNet Editing Tool". This causes confusion with L10-1180, which is also titled "GernEdiT – The GermaNet Editing Tool". I therefore propose to use the PDF's title in the metadata.

* Replaced dash with colon